feat: integration ikp code in detail aduan
